requirements.txt
django
djangorestframework
django-filter
django-rest-swagger
drf-spectacular[sidecar]
django-extensions
pyotp
django-cors-headers